Fix configurator dashboard schema import
